Complete problem 7 for C++

+++ problem-3/c-cpp/problem-3.c
@@ -10,11 +10,11 @@ int main()
     unsigned char* prime_sieve = getPrimeSieve(composite_sqrt);
 
     // Iterate sieve in reverse and mod the given composite until it equals 0
-    for (int j = composite_sqrt - 1; j > 0; j--)
+    for (long int j = composite_sqrt - 1; j > 0; j--)
     {
-        if (prime_sieve[j] != 0 && given_composite % ((long int) j) == 0)
+        if (prime_sieve[j] != 0 && given_composite % j == 0)
         {
-            printf("%d\n", j);
+            printf("%ld\n", j);
             break;
         }
     }
